requirements.txt

apache-airflow
numpy
opencv-python-headless